feat: replace PeopleTable with PeoplePage
